6/6 commit

+++ CMApp/src/main/java/CMServerEventHandler.java
@@ -18,13 +18,16 @@ public class CMServerEventHandler implements CMAppEventHandler {
 
     HashMap<String, HashMap<String, String>> shareMap; //<파일명, 공유중안 사용자> 쌍
 
+    HashMap<String, Integer> logicalClock;
+
+    String[] strServerFiles;
+
     public CMServerEventHandler(CMServerStub serverStub)
     {
         m_serverStub = serverStub;
         shareMap = new HashMap<String, HashMap<String, String>>();//{보내는사람: {파일이름:공유받을사람}}
-
+        logicalClock = new HashMap<String, Integer>();
     }
-
 
     public void processEvent(CMEvent cme)
     {
@@ -37,6 +40,9 @@ public class CMServerEventHandler implements CMAppEventHandler {
                     processSyncEvent(cme);
                 else if(cme.getID() == 106)
                     processShareEvent(cme);
+                else if(cme.getID() == 107)
+                    DeleteFile(cme);
+
             }
             default -> {
                 return;
@@ -103,8 +109,6 @@ public class CMServerEventHandler implements CMAppEventHandler {
     private void processFileEvent(CMEvent cme)
     {
         CMFileEvent fe = (CMFileEvent) cme;
-
-
         switch(fe.getID())
         {
             case CMFileEvent.REQUEST_PERMIT_PULL_FILE:
@@ -248,10 +252,12 @@ public class CMServerEventHandler implements CMAppEventHandler {
 
     private void processSyncEvent(CMEvent cme)
     {
+        //클라이언트에서 모든 파일 정보를 받아놓고, 서버에서는 보내도 되는 파일 정보만을 담아서 클라이언트에게 반환
+
         CMDummyEvent due = (CMDummyEvent) cme;
         String sender = due.getSender();
 
-        String[] strServerFiles = null;
+        //String[] strServerFiles = null;
         String[] absServerFiles = null;
         String fileServerDir = null;
         String strServerFileList = null;
@@ -263,31 +269,67 @@ public class CMServerEventHandler implements CMAppEventHandler {
         File serverDir = new File(fileServerDir);
         strServerFiles = serverDir.list();
 
+
         nServerFileNum = strServerFiles.length;
 
 
         //클라이언트에서 받은 파일 리스트 쪼개서 불러오기
         String strClientFileList = null;
         String[] strClientFiles = null;
-        int nClienFiletNum = -1;
+        int nClientFileNum = -1;
 
         strClientFileList = due.getDummyInfo();
         strClientFileList.trim();
-        strClientFiles = strClientFileList.split(" ");
+
+        String[] strClientClockFiles = null;//시계:파일이름 쌍으로 구성
+        strClientClockFiles = strClientFileList.split(" ");
+        nClientFileNum = strClientClockFiles.length;
 
 
-        List<String> strClientList = new ArrayList<>(Arrays.asList(strClientFiles));
-        for(int i = 0; i < nServerFileNum; i++)
+        //만약 보낼 파일이 없다면, nClientFileNum을 0으로 세팅
+        if(strClientFileList.equals(""))
+            nClientFileNum = 0 ;
+        System.out.println("n: " + nClientFileNum);
+        System.out.println("isEmpty: " + strClientFileList.equals(""));
+        List<String> acceptedFilesList = new ArrayList<String>();//동기화해도 되는 파일 목록
+        List<String> strClientList = new ArrayList<>();
+        //클라이언트에서 받아온 요청 메시지 파싱
+
+        for(int i = 0 ; i < nClientFileNum; i++)
         {
-            if(strClientList.contains(strServerFiles[i])==false)
+            String[] strClientClockFile = strClientClockFiles[i].split(":");
+            //System.out.println("================");
+            String clientFile = strClientClockFile[0];//동기화를 요청하고자 하는 파일 이름
+            int clientClock = Integer.parseInt(strClientClockFile[1]);//동기화를 요청하고자 하는 파일의 logical clock값
+
+            strClientList.add(clientFile);//일단 동기화를 요청한 파일을 모두 리스트에 추가
+
+            if(logicalClock.getOrDefault(clientFile, 0) < clientClock)
             {
-                //동기화 요청한 client에서 삭제된 파일은 서버에서도 삭제
-                String absServerFile = fileServerDir + "/" + strServerFiles[i];
-                System.out.println("delete: " + absServerFile);
-                File file = new File(absServerFile);
-                file.delete();
+                System.out.println(clientFile + " Sync accepted");
+                acceptedFilesList.add(clientFile);
+                //logical clock update
+                logicalClock.put(clientFile, clientClock);
+                System.out.println(clientFile + " logical clock updated: " + clientClock);
             }
             else
+            {
+                System.out.println(clientFile + " Sync rejected");
+            }
+        }
+
+        //elseif에서는 한 번 logical Clock 기준으로 필터링된 파일들만 strClientList로 만들어 주어야 함
+        //List<String> strClientList = new ArrayList<>(Arrays.asList(strClientFiles));
+
+        //따로 함수파는 게 나을듯?
+        int nAcceptedFilesNum = acceptedFilesList.size();
+        String[] acceptedClientFiles = acceptedFilesList.toArray(new String[nAcceptedFilesNum]);
+
+        for(int i = 0; i < nServerFileNum; i++)
+        {
+            //동기화가 허용되어서 다시 전송을 요청할 파일들만 삭제
+            //--------------이부분부터 다시 시작------------
+            if(acceptedFilesList.contains(strServerFiles[i]))
             {
                 //추후에 수정된 파일만 지우고 다시보내게끔 elseif로 고치면 됨!
                 String absServerFile = fileServerDir + "/" + strServerFiles[i];
@@ -295,7 +337,15 @@ public class CMServerEventHandler implements CMAppEventHandler {
                 File file = new File(absServerFile);
                 file.delete();
             }
+
         }
+        System.out.println(String.join(" ", acceptedFilesList));
+        CMDummyEvent newDue = new CMDummyEvent();
+        newDue.setID(105);
+        newDue.setDummyInfo(String.join(" ", acceptedFilesList));
+        //동기화 가능한 파일 정보를 더미에 담아서 다시 돌려줌
+        m_serverStub.send(newDue, sender);
+        strServerFiles = serverDir.list();
 
         return ;
     }
@@ -313,15 +363,31 @@ public class CMServerEventHandler implements CMAppEventHandler {
         fileName = commands[0];
         strTarget = commands[1];
 
+
+        //보내려는 클라이언트에게 사전에 파일과의 링크 생성
+        CMDummyEvent linkDue = new CMDummyEvent();
+        linkDue.setID(108);
+        linkDue.setDummyInfo(fileName + " " + sender);
+        m_serverStub.send(linkDue, strTarget);
+
+        //파일이름 + logical clock값을 실어서 보냄
+        String fileClock = fileName + ":" + logicalClock.getOrDefault(fileName, 0);
         //파일을 전송하기 전에 전송받을 파일명을 클라이언트에게 미리 알려, 업데이트된 파일을 지우도록 지시(나중에 더미메시지에 업데이트날짜도 실어보내야됨)
+
         CMDummyEvent newDue = new CMDummyEvent();
         newDue.setID(106);
-        newDue.setDummyInfo(fileName);
+        //newDue.setDummyInfo(fileName);
+        newDue.setDummyInfo(fileClock);
         m_serverStub.send(newDue, strTarget);
 
         strFile = "./server-file-path/" + sender + "/" + fileName;//
         System.out.println(strFile);
 
+        try {
+            Thread.sleep(1000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
         boolean bReturn = false;
         bReturn = CMFileTransferManager.pushFile(strFile, strTarget, m_serverStub.getCMInfo());
         //bReturn = m_serverStub.pushFile(strFile, strTarget, CMInfo.FILE_OVERWRITE);
@@ -332,13 +398,38 @@ public class CMServerEventHandler implements CMAppEventHandler {
         return ;
     }
 
-    void DeleteFile(String strFile)
+    void DeleteFile(CMEvent cme)
     {
+        CMDummyEvent due = (CMDummyEvent) cme;
+        String sender = due.getSender();
+        String strFilesMessage = null;
+        String[] strFiles = null;
+        String[] strServerFiles = null;
+        strFilesMessage = due.getDummyInfo();
+        strFilesMessage.trim();
+
         String fileClientDir = "./server-file-path";
-        String absServerFile = fileClientDir + "/" + strFile;
-        System.out.println("delete: " + absServerFile);
-        File file = new File(absServerFile);
-        file.delete();
+        String absServerDir = fileClientDir + "/" + sender + "/";
+        File serverFiles = new File(absServerDir);
+        strServerFiles = serverFiles.list();
+
+        strFiles = strFilesMessage.split(" ");
+        int nFileNum = strServerFiles.length;
+        System.out.println("nFileNum: " + nFileNum);
+
+        //클라에서 삭제된 파일을 서버에서도 삭제
+        for(int i = 0; i < nFileNum; i++)
+        {
+            List strFilesList = Arrays.asList(strFiles);
+            String absServerFile = absServerDir + strServerFiles[i];
+            //System.out.println("file: " + absServerFile);
+            if(strFilesList.contains(strServerFiles[i]) == false)
+            {
+                System.out.println("delete: " + absServerFile);
+                File file = new File(absServerFile);
+                file.delete();
+            }
+        }
     }
 
 }
